fixup startup-check for fee account view key
In the verification procedure, `TxOut::view_key_match` function is
used to test if a TxOut belongs to the fee account. However, this
will not work with the subaddress view key, it needs the account
view key. So the check placed in server/src/bin/main.rs was actually
wrong -- that check only passes for the subaddress viwe key, which
is actually the wrong key.